A major rewrite of the data transfer between correlator and input node. The data is no longer broken into packets at the input node but is send as one big stream. The integer delay now also is handled at the correlator node. The major benefit is that we can now have an arbitrary number of lags(fft window) in a simple and transparent way.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@1005 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/input_node_data_writer_tasklet.h
@@ -77,10 +77,10 @@ public:
   * @desc Add a pair of (data_writer, size_slice) to a given channel-queue.
   * @param int nr_stream The identifier of the stream.
   * @param Data_writer_sptr wr The writer on which to stream the data
-  * @param int size the amount of byte to send to the given writer
+  * @param int64_t size the amount of samples to send to the given writer
   * assert( nr_stream < number_channels() )
   *****************************************************************************/
-  void add_timeslice_to_stream(int nr_stream, Data_writer_sptr wr, int size);
+  void add_timeslice_to_stream(int nr_stream, Data_writer_sptr wr, int64_t size);
 
   /*****************************************************************************
   * @desc Deplate the input_queues of all of the data_writers.
